Basic object interaction
Objects in rooms can be picked up which removes them from the room and adds them to player inventory. Objects can be looked at which give their description. objects can be dropped from the inventory into the room the player is in. the inventory can be checked now as well.